Uses errors.Is API to check wrapped errors for their type
errors.Is is the preferred API for checking errors as it will
recursively unwrap error objects to find a match. In this case, we were
wrapping an error in the parser which then meant that the conditional
was no longer triggered and resulted in a stray error instead of a
failure when running detection.